Integrate LRN/LRNGrad op with MIOpen
Integrate LRN/LRNGrad op with MIOpen. NHWC<->NCHW conversion is carried out to
make sure only NCHW tensors are passed to MIOpen for computation, as it's the
only format supported now.

The implementation so far assumes input/output tensors from TensorFlow are
always in NHWC format. This should be further improved to consider cases when
the inputs/outputs are already in NCHW format so conversions can be skipped.